+++ benight-admin/src/app/core/interfaces/interfaces.languages.ts
@@ -1,8 +1,10 @@
 export namespace languages {
     
-    export const es = 'es_ES'
-    export const fr = 'fr_FR'
-    export const pt = 'pt_PT'
-    export const en = 'en_EN' 
+    export const es = 'es-ES'
+    export const fr = 'fr-FR'
+    export const pt = 'pt-PT'
+    export const en = 'en-EN'
+    
+    export const language_list = ['Es','En','Fr','Pt']
   
   }