+++ benight-admin/src/app/core/interfaces/interfaces.modal.ts
@@ -1,0 +1,7 @@
+export namespace modals {
+    //Modals
+    export const MD_EVENT = 0
+    export const MD_PLAN = 1
+    export const MD_CLUB = 2
+    export const MD_TICKET = 3
+}