+++ benight-admin/src/app/modals/clubs/preview-club.page.ts
@@ -1,0 +1,13 @@
+import { Component, OnInit } from '@angular/core'
+
+@Component({
+  selector: 'preview-club',
+  templateUrl: './preview-club.page.html',
+  styleUrls: ['./preview-club.page.scss'],
+})
+export class PreviewClubPage implements OnInit {
+
+  constructor() { }
+
+  ngOnInit() {}
+}