+++ benight-admin/src/app/modals/entrada/preview-entrada.page.ts
@@ -1,0 +1,13 @@
+import { Component, OnInit } from '@angular/core'
+
+@Component({
+  selector: 'preview-entrada',
+  templateUrl: './preview-entrada.page.html',
+  styleUrls: ['./preview-entrada.page.scss'],
+})
+export class PreviewEntradaPage implements OnInit {
+
+  constructor() { }
+
+  ngOnInit() {}
+}