+++ benight-admin/src/app/modals/planes/preview-planes.page.ts
@@ -1,0 +1,13 @@
+import { Component, OnInit } from '@angular/core'
+
+@Component({
+  selector: 'preview-planes',
+  templateUrl: './preview-planes.page.html',
+  styleUrls: ['./preview-planes.page.scss'],
+})
+export class PreviewPlanesPage implements OnInit {
+
+  constructor() { }
+
+  ngOnInit() {}
+}