+++ benight-admin/src/app/navigator/detalle/detalle-info-event/detalle-info-event.module.ts
@@ -2,8 +2,8 @@ import { NgModule } from '@angular/core'
 
 import { myBaseModules } from '@bn8-imports/imports.main'
 import { DetalleInfoEventPage } from '@bn8-imports/imports.views'
+import { PreviewEventosPage } from '@bn8-imports/imports.previews'
 import { DetalleInfoEventRoutingModule } from '@bn8-imports/imports.routing'
-
 import { ReactiveFormsModule } from '@angular/forms'
 
 @NgModule({
@@ -12,6 +12,6 @@ import { ReactiveFormsModule } from '@angular/forms'
     ReactiveFormsModule,
     DetalleInfoEventRoutingModule
   ],
-  declarations: [DetalleInfoEventPage]
+  declarations: [DetalleInfoEventPage, PreviewEventosPage]
 })
 export class DetalleInfoEventPageModule {}