+++ benight-admin/src/app/navigator/detalle/detalle-info-event/detalle-info-event.page.ts
@@ -1,5 +1,7 @@
 import { Component, OnInit } from '@angular/core'
 import { FormBuilder, FormGroup } from '@angular/forms'
+import { ModalController } from '@ionic/angular'
+import { PreviewEventosPage } from '@bn8-imports/imports.previews' 
 
 @Component({
   selector: 'detalle-detalle-info-event',
@@ -12,7 +14,7 @@ export class DetalleInfoEventPage implements OnInit {
   create: boolean = true
   minDate: string
 
-  constructor(private fb: FormBuilder) { }
+  constructor(private fb: FormBuilder, public modal: ModalController) { }
 
   ngOnInit() {
     const actualDate: string = new Date().toISOString()
@@ -40,5 +42,11 @@ export class DetalleInfoEventPage implements OnInit {
     })
 
     this.myForm.valueChanges.subscribe(console.log)
+  }
+
+  async presentModal() {
+    const window = await this.modal.create({component: PreviewEventosPage })
+    return await window.present()
+
   }
 }