+++ benight-admin/src/app/navigator/gestion/eventos/eventos.module.ts
@@ -1,7 +1,8 @@
 import { NgModule } from '@angular/core'
 
 import { myBaseModules } from '@bn8-imports/imports.main'
-import { EventosPage,DetalleInfoEventPage } from '@bn8-imports/imports.views'
+import { EventosPage, DetalleInfoEventPage } from '@bn8-imports/imports.views'
+import { PreviewEventosPage } from '@bn8-imports/imports.previews'
 import { EventosRoutingModule} from '@bn8-imports/imports.routing'
 import { ReactiveFormsModule } from '@angular/forms'
 
@@ -11,6 +12,7 @@ import { ReactiveFormsModule } from '@angular/forms'
     ReactiveFormsModule,
     EventosRoutingModule
   ],
-  declarations: [EventosPage,DetalleInfoEventPage]
+  entryComponents: [PreviewEventosPage],
+  declarations: [EventosPage,DetalleInfoEventPage,PreviewEventosPage]
 })
 export class EventosPageModule {}